Split project into engine lib and app executable targets. Fixed dll copy for all generators. Now building a target tree for VS and XCode solutions.

+++ Code/Rendering/FrameStats.cpp
@@ -18,7 +18,7 @@ namespace te
 		if (frameTime - lastUpdate >= 1000)
 		{
 			lastUpdate = newTime;
-			fps = currFps;
+			fps = static_cast<uint64_t>(currFps);
 		}
 		frameTime = newTime;
 	}
